fix: resolve search box initialization failure on Turbo navigation
- Address page initialization errors by applying retry and defensive mechanisms
- Add orphaned DOM element cleanup to prevent layout interference
- Refactor search modules
